requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.0.2
oauthlib==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-discogs-client==2.8
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
